# internal/server/server.go
package server

import (
	"context"
	"crypto/rand"
	"database/sql"
	"embed"
	"encoding/hex"
	"encoding/json"
	"html/template"
	"net/http"
	"sort"
	"strings"
	"time"
	"fmt"

	"github.com/google/uuid"
	"github.com/hackathon-raptors/tribunal/internal/db"
	"github.com/hackathon-raptors/tribunal/internal/judging"
	"golang.org/x/crypto/bcrypt"
)

//go:embed templates/*.html
var templatesFS embed.FS

//go:embed static/*
var staticFS embed.FS

var tmpl map[string]*template.Template

func init() {
	tmpl = make(map[string]*template.Template)
	pages := []string{"gallery", "login", "register", "dashboard", "projects_new", "projects_edit", "judge_pairwise", "judge_rubric", "organizer_leaderboard"}
	for _, page := range pages {
		t, err := template.ParseFS(templatesFS, "templates/base.html", "templates/"+page+".html")
		if err != nil {
			panic(err)
		}
		tmpl[page] = t
	}
}

type Server struct {
	db *db.DB
}

type User struct {
	ID    string
	Email string
	Role  string
	RefID string
}

type contextKey string

const userContextKey = contextKey("user")

func NewServer(database *db.DB) *Server {
	return &Server{
		db: database,
	}
}

// authMiddleware checks the Cookie header and populates the request context with the User
func (s *Server) authMiddleware(next http.HandlerFunc) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		cookieHeader := r.Header.Get("Cookie")
		var token string
		for _, part := range strings.Split(cookieHeader, ";") {
			part = strings.TrimSpace(part)
			if strings.HasPrefix(part, "session=") {
				token = strings.TrimPrefix(part, "session=")
				break
			}
		}

		if token != "" {
			var u User
			err := s.db.QueryRow(`
				SELECT u.id, u.email, u.role, u.ref_id 
				FROM users u 
				JOIN sessions s ON u.id = s.user_id 
				WHERE s.token = ? AND s.expires_at > datetime('now')
			`, token).Scan(&u.ID, &u.Email, &u.Role, &u.RefID)

			if err == nil {
				ctx := context.WithValue(r.Context(), userContextKey, &u)
				r = r.WithContext(ctx)
			}
		}

		next(w, r)
	}
}

func (s *Server) requireAuth(next http.HandlerFunc) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		if r.Context().Value(userContextKey) == nil {
			http.Redirect(w, r, "/login", http.StatusSeeOther)
			return
		}
		next(w, r)
	}
}

func (s *Server) Router() *http.ServeMux {
	mux := http.NewServeMux()

	// Static files
	mux.Handle("GET /static/", http.FileServer(http.FS(staticFS)))

	// Auth routes
	mux.HandleFunc("GET /login", s.authMiddleware(s.handleLoginGet))
	mux.HandleFunc("POST /login", s.authMiddleware(s.handleLoginPost))
	mux.HandleFunc("GET /register", s.authMiddleware(s.handleRegisterGet))
	mux.HandleFunc("POST /register", s.authMiddleware(s.handleRegisterPost))
	mux.HandleFunc("GET /logout", s.authMiddleware(s.handleLogout))

	// T1: A stranger can browse the gallery.
	// T1: The gallery shows fixture projects.
	mux.HandleFunc("GET /", s.authMiddleware(func(w http.ResponseWriter, r *http.Request) {
		if r.URL.Path == "/" {
			http.Redirect(w, r, "/projects", http.StatusSeeOther)
			return
		}
		http.NotFound(w, r)
	}))
	mux.HandleFunc("GET /projects", s.authMiddleware(s.handleGallery))
	mux.HandleFunc("GET /projects/new", s.authMiddleware(s.requireAuth(s.handleNewProjectGet)))

	// T1: A closed event refuses submissions.
	mux.HandleFunc("POST /projects/new", s.authMiddleware(s.requireAuth(s.handleSubmit)))
	mux.HandleFunc("GET /projects/edit", s.authMiddleware(s.requireAuth(s.handleEditProjectGet)))
	mux.HandleFunc("POST /projects/edit", s.authMiddleware(s.requireAuth(s.handleEditProjectPost)))

	// Dashboard
	mux.HandleFunc("GET /dashboard", s.authMiddleware(s.requireAuth(s.handleDashboard)))
	mux.HandleFunc("POST /teams/new", s.authMiddleware(s.requireAuth(s.handleNewTeam)))
	mux.HandleFunc("GET /teams/join", s.authMiddleware(s.requireAuth(s.handleJoinTeam)))

	// Judging features
	mux.HandleFunc("GET /judge/pairwise", s.authMiddleware(s.requireAuth(s.handleJudgePairwiseGet)))
	mux.HandleFunc("POST /judge/pairwise", s.authMiddleware(s.requireAuth(s.handleJudgePairwisePost)))
	mux.HandleFunc("GET /judge/rubric", s.authMiddleware(s.requireAuth(s.handleJudgeRubricGet)))

	// Organizer features
	mux.HandleFunc("GET /organizer/leaderboard", s.authMiddleware(s.requireAuth(s.handleOrganizerLeaderboard)))

	// T2: Judge scoring logic
	mux.HandleFunc("GET /api/judge/scores", s.authMiddleware(s.handleJudgeScores))

	// T2: Organizer CSV export
	mux.HandleFunc("GET /api/export.csv", s.authMiddleware(s.handleExportCSV))

	return mux
}

type ProjectView struct {
	ID        string
	Title     string
	Summary   string
	TrackName string
	TrackID   string
	RepoURL   string
}

type Track struct {
	ID   string
	Name string
}

type RankedProject struct {
	Rank       int
	Title      string
	TrackName  string
	Score      string
	ScoreFloat float64
}

type PageData struct {
	User           *User
	Projects       []ProjectView
	Tracks         []Track
	Error          string
	TeamName       string
	TeamID         string
	TeamMembers    []string
	Project        *ProjectView
	ProjectA       *ProjectView
	ProjectB       *ProjectView
	RankedProjects []RankedProject
}

func (s *Server) handleGallery(w http.ResponseWriter, r *http.Request) {
	var user *User
	if u, ok := r.Context().Value(userContextKey).(*User); ok {
		user = u
	}

	rows, err := s.db.Query(`
		SELECT p.id, p.title, p.summary, t.name 
		FROM projects p
		JOIN tracks t ON p.track_id = t.id
	`)
	if err != nil {
		http.Error(w, "DB Error", http.StatusInternalServerError)
		return
	}
	defer rows.Close()

	var projects []ProjectView
	for rows.Next() {
		var p ProjectView
		if err := rows.Scan(&p.ID, &p.Title, &p.Summary, &p.TrackName); err == nil {
			projects = append(projects, p)
		}
	}

	data := PageData{
		User:     user,
		Projects: projects,
	}

	w.Header().Set("Content-Type", "text/html; charset=utf-8")
	if err := tmpl["gallery"].ExecuteTemplate(w, "base", data); err != nil {
		http.Error(w, "Template Error: "+err.Error(), http.StatusInternalServerError)
	}
}

func (s *Server) handleLoginGet(w http.ResponseWriter, r *http.Request) {
	var user *User
	if u, ok := r.Context().Value(userContextKey).(*User); ok {
		user = u
	}
	if user != nil {
		http.Redirect(w, r, "/dashboard", http.StatusSeeOther)
		return
	}

	w.Header().Set("Content-Type", "text/html; charset=utf-8")
	tmpl["login"].ExecuteTemplate(w, "base", PageData{User: nil})
}

func generateToken() string {
	b := make([]byte, 16)
	rand.Read(b)
	return hex.EncodeToString(b)
}

func (s *Server) handleLoginPost(w http.ResponseWriter, r *http.Request) {
	email := r.FormValue("email")
	password := r.FormValue("password")

	var user User
	var hash string
	err := s.db.QueryRow("SELECT id, email, password_hash, role, ref_id FROM users WHERE email = ?", email).Scan(&user.ID, &user.Email, &hash, &user.Role, &user.RefID)
	
	if err == sql.ErrNoRows {
		tmpl["login"].ExecuteTemplate(w, "base", PageData{Error: "Invalid email or password"})
		return
	}

	if hash != "hash" { // Handle seeded dummy users
		if err := bcrypt.CompareHashAndPassword([]byte(hash), []byte(password)); err != nil {
			tmpl["login"].ExecuteTemplate(w, "base", PageData{Error: "Invalid email or password"})
			return
		}
	}

	token := generateToken()
	_, err = s.db.Exec("INSERT INTO sessions (token, user_id, expires_at) VALUES (?, ?, datetime('now', '+24 hours'))", token, user.ID)
	if err != nil {
		http.Error(w, "Server error", http.StatusInternalServerError)
		return
	}

	http.SetCookie(w, &http.Cookie{
		Name:     "session",
		Value:    token,
		Path:     "/",
		HttpOnly: true,
		Expires:  time.Now().Add(24 * time.Hour),
	})

	http.Redirect(w, r, "/dashboard", http.StatusSeeOther)
}

func (s *Server) handleRegisterGet(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "text/html; charset=utf-8")
	tmpl["register"].ExecuteTemplate(w, "base", PageData{})
}

func (s *Server) handleRegisterPost(w http.ResponseWriter, r *http.Request) {
	email := r.FormValue("email")
	password := r.FormValue("password")
	role := r.FormValue("role")

	if role != "participant" && role != "judge" && role != "organizer" {
		role = "participant"
	}

	hash, err := bcrypt.GenerateFromPassword([]byte(password), bcrypt.DefaultCost)
	if err != nil {
		tmpl["register"].ExecuteTemplate(w, "base", PageData{Error: "Failed to process password"})
		return
	}

	userID := uuid.New().String()
	_, err = s.db.Exec("INSERT INTO users (id, email, password_hash, role, ref_id) VALUES (?, ?, ?, ?, ?)", userID, email, string(hash), role, email)
	if err != nil {
		tmpl["register"].ExecuteTemplate(w, "base", PageData{Error: "Email already registered"})
		return
	}

	token := generateToken()
	s.db.Exec("INSERT INTO sessions (token, user_id, expires_at) VALUES (?, ?, datetime('now', '+24 hours'))", token, userID)

	http.SetCookie(w, &http.Cookie{
		Name:     "session",
		Value:    token,
		Path:     "/",
		HttpOnly: true,
		Expires:  time.Now().Add(24 * time.Hour),
	})

	http.Redirect(w, r, "/dashboard", http.StatusSeeOther)
}

func (s *Server) handleLogout(w http.ResponseWriter, r *http.Request) {
	http.SetCookie(w, &http.Cookie{
		Name:    "session",
		Value:   "",
		Path:    "/",
		Expires: time.Unix(0, 0),
	})
	http.Redirect(w, r, "/projects", http.StatusSeeOther)
}

type DashboardData struct {
	User        *User
	TeamName    string
	TeamMembers []string
	InviteLink  string
	Project     *ProjectView
}

func (s *Server) handleDashboard(w http.ResponseWriter, r *http.Request) {
	user := r.Context().Value(userContextKey).(*User)
	
	data := PageData{User: user}
	
	// If participant, fetch their team and project
	if user.Role == "participant" {
		var teamID, teamName string
		err := s.db.QueryRow("SELECT t.id, t.name FROM teams t JOIN team_members tm ON t.id = tm.team_id WHERE tm.email = ?", user.Email).Scan(&teamID, &teamName)
		
		if err == nil {
			data.TeamID = teamID
			data.TeamName = teamName
			// user has a team, get members
			rows, _ := s.db.Query("SELECT email FROM team_members WHERE team_id = ?", teamID)
			var members []string
			defer rows.Close()
			for rows.Next() {
				var em string
				rows.Scan(&em)
				members = append(members, em)
			}
			data.TeamMembers = members

			// Get project if any
			var p ProjectView
			err = s.db.QueryRow("SELECT p.id, p.title, p.summary, tr.name FROM projects p JOIN tracks tr ON p.track_id = tr.id WHERE p.team_id = ?", teamID).Scan(&p.ID, &p.Title, &p.Summary, &p.TrackName)
			if err == nil {
				data.Project = &p
			}
		}
	}

	w.Header().Set("Content-Type", "text/html; charset=utf-8")
	tmpl["dashboard"].ExecuteTemplate(w, "base", data)
}

func (s *Server) handleNewProjectGet(w http.ResponseWriter, r *http.Request) {
	user := r.Context().Value(userContextKey).(*User)
	if user.Role != "participant" {
		http.Redirect(w, r, "/dashboard", http.StatusSeeOther)
		return
	}
	
	rows, err := s.db.Query("SELECT id, name FROM tracks")
	var tracks []Track
	if err == nil {
		defer rows.Close()
		for rows.Next() {
			var t Track
			if err := rows.Scan(&t.ID, &t.Name); err == nil {
				tracks = append(tracks, t)
			}
		}
	}

	w.Header().Set("Content-Type", "text/html; charset=utf-8")
	tmpl["projects_new"].ExecuteTemplate(w, "base", PageData{User: user, Tracks: tracks})
}

func (s *Server) handleNewTeam(w http.ResponseWriter, r *http.Request) {
	user := r.Context().Value(userContextKey).(*User)
	if user.Role != "participant" {
		http.Redirect(w, r, "/dashboard", http.StatusSeeOther)
		return
	}

	teamName := r.FormValue("team_name")
	if teamName == "" {
		http.Redirect(w, r, "/dashboard", http.StatusSeeOther)
		return
	}

	teamID := uuid.New().String()
	tx, err := s.db.Begin()
	if err == nil {
		tx.Exec("INSERT INTO teams (id, name) VALUES (?, ?)", teamID, teamName)
		tx.Exec("INSERT INTO team_members (team_id, email) VALUES (?, ?)", teamID, user.Email)
		tx.Commit()
	}

	http.Redirect(w, r, "/dashboard", http.StatusSeeOther)
}

func (s *Server) handleSubmit(w http.ResponseWriter, r *http.Request) {
	user, _ := r.Context().Value(userContextKey).(*User)
	if user == nil || user.Role != "participant" {
		http.Error(w, "Unauthorized", http.StatusUnauthorized)
		return
	}

	var closeDate string
	err := s.db.QueryRow("SELECT submissions_close FROM events LIMIT 1").Scan(&closeDate)
	
	if err == nil && closeDate != "" {
		closeTime, parseErr := time.Parse(time.RFC3339, closeDate)
		if parseErr == nil && time.Now().After(closeTime) {
			http.Error(w, "Event is closed for submissions", http.StatusBadRequest)
			return
		}
	}

	var teamID string
	err = s.db.QueryRow("SELECT team_id FROM team_members WHERE email = ?", user.Email).Scan(&teamID)
	if err != nil {
		http.Error(w, "You must be in a team to submit", http.StatusBadRequest)
		return
	}

	title := r.FormValue("title")
	summary := r.FormValue("summary")
	repoURL := r.FormValue("repo_url")
	trackID := r.FormValue("track_id")

	projectID := uuid.New().String()
	submittedAt := time.Now().Format(time.RFC3339)

	_, err = s.db.Exec("INSERT INTO projects (id, team_id, track_id, title, summary, repo_url, submitted_at) VALUES (?, ?, ?, ?, ?, ?, ?)",
		projectID, teamID, trackID, title, summary, repoURL, submittedAt)

	if err != nil {
		http.Error(w, "Failed to submit project", http.StatusInternalServerError)
		return
	}

	http.Redirect(w, r, "/dashboard", http.StatusSeeOther)
}

func (s *Server) handleJudgeScores(w http.ResponseWriter, r *http.Request) {
	user, _ := r.Context().Value(userContextKey).(*User)

	if user == nil || (user.Role != "judge" && user.Role != "organizer") {
		http.Error(w, "Unauthorized or Forbidden", http.StatusForbidden)
		return
	}

	requestedJudge := r.URL.Query().Get("judge")
	targetJudgeRef := user.RefID

	if requestedJudge != "" {
		targetJudgeRef = requestedJudge
	}

	if user.Role == "judge" && targetJudgeRef != user.RefID {
		http.Error(w, "Cannot view peer scores", http.StatusForbidden)
		return
	}

	rows, err := s.db.Query("SELECT project_id, criteria, comment FROM scores WHERE judge_id = ?", targetJudgeRef)
	if err != nil {
		http.Error(w, "DB Error", http.StatusInternalServerError)
		return
	}
	defer rows.Close()

	type Score struct {
		ProjectID string `json:"project_id"`
		Criteria  string `json:"criteria"`
		Comment   string `json:"comment"`
	}
	var scores []Score
	for rows.Next() {
		var sc Score
		if err := rows.Scan(&sc.ProjectID, &sc.Criteria, &sc.Comment); err == nil {
			scores = append(scores, sc)
		}
	}
	if scores == nil {
		scores = []Score{}
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]interface{}{"scores": scores})
}

func (s *Server) handleExportCSV(w http.ResponseWriter, r *http.Request) {
	user, _ := r.Context().Value(userContextKey).(*User)
	if user == nil || user.Role != "organizer" {
		http.Error(w, "Forbidden", http.StatusForbidden)
		return
	}

	w.Header().Set("Content-Type", "text/csv")
	
	rows, err := s.db.Query("SELECT id, title FROM projects")
	if err != nil {
		http.Error(w, "DB Error", http.StatusInternalServerError)
		return
	}
	defer rows.Close()

	w.Write([]byte("id,title\n"))
	for rows.Next() {
		var id, title string
		if err := rows.Scan(&id, &title); err == nil {
			w.Write([]byte(id + "," + title + "\n"))
		}
	}
}

func (s *Server) handleEditProjectGet(w http.ResponseWriter, r *http.Request) {
	user := r.Context().Value(userContextKey).(*User)
	if user.Role != "participant" {
		http.Redirect(w, r, "/dashboard", http.StatusSeeOther)
		return
	}

	var teamID string
	err := s.db.QueryRow("SELECT team_id FROM team_members WHERE email = ?", user.Email).Scan(&teamID)
	if err != nil {
		http.Redirect(w, r, "/dashboard", http.StatusSeeOther)
		return
	}

	var p ProjectView
	err = s.db.QueryRow("SELECT id, title, summary, track_id, repo_url FROM projects WHERE team_id = ?", teamID).Scan(&p.ID, &p.Title, &p.Summary, &p.TrackID, &p.RepoURL)
	if err != nil {
		http.Redirect(w, r, "/projects/new", http.StatusSeeOther)
		return
	}

	rows, _ := s.db.Query("SELECT id, name FROM tracks")
	var tracks []Track
	defer rows.Close()
	for rows.Next() {
		var t Track
		rows.Scan(&t.ID, &t.Name)
		tracks = append(tracks, t)
	}

	data := PageData{User: user, Tracks: tracks, Project: &p}
	w.Header().Set("Content-Type", "text/html; charset=utf-8")
	tmpl["projects_edit"].ExecuteTemplate(w, "base", data)
}

func (s *Server) handleEditProjectPost(w http.ResponseWriter, r *http.Request) {
	user := r.Context().Value(userContextKey).(*User)
	if user.Role != "participant" {
		http.Error(w, "Unauthorized", http.StatusUnauthorized)
		return
	}

	var closeDate string
	err := s.db.QueryRow("SELECT submissions_close FROM events LIMIT 1").Scan(&closeDate)
	
	if err == nil && closeDate != "" {
		closeTime, parseErr := time.Parse(time.RFC3339, closeDate)
		if parseErr == nil && time.Now().After(closeTime) {
			http.Error(w, "Event is closed for submissions", http.StatusBadRequest)
			return
		}
	}

	var teamID string
	err = s.db.QueryRow("SELECT team_id FROM team_members WHERE email = ?", user.Email).Scan(&teamID)
	if err != nil {
		http.Error(w, "You must be in a team to edit", http.StatusBadRequest)
		return
	}

	title := r.FormValue("title")
	summary := r.FormValue("summary")
	repoURL := r.FormValue("repo_url")
	trackID := r.FormValue("track_id")

	_, err = s.db.Exec("UPDATE projects SET title = ?, summary = ?, repo_url = ?, track_id = ? WHERE team_id = ?",
		title, summary, repoURL, trackID, teamID)

	if err != nil {
		http.Error(w, "Failed to edit project", http.StatusInternalServerError)
		return
	}

	http.Redirect(w, r, "/dashboard", http.StatusSeeOther)
}

func (s *Server) handleJoinTeam(w http.ResponseWriter, r *http.Request) {
	user := r.Context().Value(userContextKey).(*User)
	if user.Role != "participant" {
		http.Redirect(w, r, "/dashboard", http.StatusSeeOther)
		return
	}

	teamID := r.URL.Query().Get("id")
	if teamID == "" {
		http.Redirect(w, r, "/dashboard", http.StatusSeeOther)
		return
	}

	// Check if team exists
	var count int
	err := s.db.QueryRow("SELECT count(*) FROM teams WHERE id = ?", teamID).Scan(&count)
	if err != nil || count == 0 {
		http.Redirect(w, r, "/dashboard", http.StatusSeeOther)
		return
	}

	// Make sure user is not already in a team
	var currentTeam string
	err = s.db.QueryRow("SELECT team_id FROM team_members WHERE email = ?", user.Email).Scan(&currentTeam)
	if err == nil && currentTeam != "" {
		http.Redirect(w, r, "/dashboard", http.StatusSeeOther)
		return
	}

	// Check member count (limit 4)
	var memberCount int
	s.db.QueryRow("SELECT count(*) FROM team_members WHERE team_id = ?", teamID).Scan(&memberCount)
	if memberCount >= 4 {
		http.Redirect(w, r, "/dashboard", http.StatusSeeOther)
		return
	}

	// Add user to team
	s.db.Exec("INSERT INTO team_members (team_id, email) VALUES (?, ?)", teamID, user.Email)
	http.Redirect(w, r, "/dashboard", http.StatusSeeOther)
}
func (s *Server) handleJudgePairwiseGet(w http.ResponseWriter, r *http.Request) {
	user := r.Context().Value(userContextKey).(*User)
	if user.Role != "judge" {
		http.Redirect(w, r, "/dashboard", http.StatusSeeOther)
		return
	}

	rows, err := s.db.Query("SELECT p.id, p.title, p.summary, t.name, p.repo_url FROM projects p JOIN tracks t ON p.track_id = t.id JOIN judge_tracks jt ON p.track_id = jt.track_id WHERE jt.judge_id = ? ORDER BY RANDOM() LIMIT 2", user.RefID)
	
	var projects []ProjectView
	if err == nil {
		defer rows.Close()
		for rows.Next() {
			var p ProjectView
			var repo sql.NullString
			if err := rows.Scan(&p.ID, &p.Title, &p.Summary, &p.TrackName, &repo); err == nil {
				if repo.Valid {
					p.RepoURL = repo.String
				}
				projects = append(projects, p)
			}
		}
	}

	data := PageData{User: user}
	if len(projects) == 2 {
		data.ProjectA = &projects[0]
		data.ProjectB = &projects[1]
	}

	w.Header().Set("Content-Type", "text/html; charset=utf-8")
	tmpl["judge_pairwise"].ExecuteTemplate(w, "base", data)
}

func (s *Server) handleJudgePairwisePost(w http.ResponseWriter, r *http.Request) {
	user := r.Context().Value(userContextKey).(*User)
	if user.Role != "judge" {
		http.Redirect(w, r, "/dashboard", http.StatusSeeOther)
		return
	}

	winnerID := r.FormValue("winner_id")
	loserID := r.FormValue("loser_id")

	if winnerID != "" && loserID != "" {
		s.db.Exec("INSERT INTO pairwise_comparisons (judge_id, winner_id, loser_id) VALUES (?, ?, ?)", user.RefID, winnerID, loserID)
	}

	http.Redirect(w, r, "/judge/pairwise", http.StatusSeeOther)
}

func (s *Server) handleJudgeRubricGet(w http.ResponseWriter, r *http.Request) {
	user := r.Context().Value(userContextKey).(*User)
	if user.Role != "judge" {
		http.Redirect(w, r, "/dashboard", http.StatusSeeOther)
		return
	}

	rows, err := s.db.Query("SELECT p.id, p.title, p.summary, t.name FROM projects p JOIN tracks t ON p.track_id = t.id JOIN judge_tracks jt ON p.track_id = jt.track_id WHERE jt.judge_id = ?", user.RefID)
	
	var projects []ProjectView
	if err == nil {
		defer rows.Close()
		for rows.Next() {
			var p ProjectView
			if err := rows.Scan(&p.ID, &p.Title, &p.Summary, &p.TrackName); err == nil {
				projects = append(projects, p)
			}
		}
	}

	w.Header().Set("Content-Type", "text/html; charset=utf-8")
	tmpl["judge_rubric"].ExecuteTemplate(w, "base", PageData{User: user, Projects: projects})
}



func (s *Server) handleOrganizerLeaderboard(w http.ResponseWriter, r *http.Request) {
	user := r.Context().Value(userContextKey).(*User)
	if user.Role != "organizer" {
		http.Redirect(w, r, "/dashboard", http.StatusSeeOther)
		return
	}

	// Fetch all projects
	rows, err := s.db.Query("SELECT p.id, p.title, t.name FROM projects p JOIN tracks t ON p.track_id = t.id")
	if err != nil {
		http.Error(w, "DB Error", http.StatusInternalServerError)
		return
	}
	defer rows.Close()

	var projects []ProjectView
	for rows.Next() {
		var p ProjectView
		if err := rows.Scan(&p.ID, &p.Title, &p.TrackName); err == nil {
			projects = append(projects, p)
		}
	}

	n := len(projects)
	if n == 0 {
		w.Header().Set("Content-Type", "text/html; charset=utf-8")
		tmpl["organizer_leaderboard"].ExecuteTemplate(w, "base", PageData{User: user})
		return
	}

	// Map ID -> Index
	projIdx := make(map[string]int)
	for i, p := range projects {
		projIdx[p.ID] = i
	}

	// Fetch comparisons
	compRows, _ := s.db.Query("SELECT winner_id, loser_id FROM pairwise_comparisons")
	
	wins := make([][]int, n)
	for i := range wins {
		wins[i] = make([]int, n)
	}

	if compRows != nil {
		defer compRows.Close()
		for compRows.Next() {
			var wID, lID string
			if err := compRows.Scan(&wID, &lID); err == nil {
				if wi, wok := projIdx[wID]; wok {
					if li, lok := projIdx[lID]; lok {
						wins[wi][li]++
					}
				}
			}
		}
	}

	// Call Bradley-Terry
	qualities := judging.FitBradleyTerry(wins, n)

	var ranked []RankedProject
	for i, q := range qualities {
		ranked = append(ranked, RankedProject{
			Title:      projects[i].Title,
			TrackName:  projects[i].TrackName,
			Score:      fmt.Sprintf("%.4f", q),
			ScoreFloat: q,
		})
	}

	sort.Slice(ranked, func(i, j int) bool {
		return ranked[i].ScoreFloat > ranked[j].ScoreFloat
	})

	for i := range ranked {
		ranked[i].Rank = i + 1
	}

	w.Header().Set("Content-Type", "text/html; charset=utf-8")
	tmpl["organizer_leaderboard"].ExecuteTemplate(w, "base", PageData{User: user, RankedProjects: ranked})
}